Move `LLVMCompiledTrace` to the jitc_llvm backend.
